Attempt to fix issue with too much nodes
There is an issue maybe on sort or the container itself as float which cause the inserted node to be duplicate even it is not existing yet

+++ gSeer/Forward Propagation/FPropagateST.cs
@@ -96,14 +96,14 @@ namespace gSeer.Forward_Propagation {
 
 	public class CalcNode : IComparable<CalcNode> {
 		ActivationFunctions.Activation ActivationFunction;
-		public float X { get; set; }
+		public double X { get; set; }
 		public float Output { get; set; }
 		public List<CalcConnection> Connections { get; set; }
 		CalcNode() {
 			ActivationFunction = new ActivationFunctions.Logistic();
 			Connections = new List<CalcConnection>();
 		}
-		public CalcNode(float x) : this() {
+		public CalcNode(double x) : this() {
 			X = x;
 		}
 		public void Calculate() {
